refactor logging to use QMods' Logger
The Logger class in QMods allows for optional display of messages in-game
as well as based on if debugging is enabled or not.
I've opted to display some messages in a pretty format in-game and keep
some to the log file only.

This should hopefully help further debug/develop ths mod.
(I've also added a few TODO:'s for translation of messages)