rotate autosaves per slot from .0000 to .9999
Each ^slot\.[0-9]{4}$ will have a number from 0000 to 9999 suffixed and
increased with every autosave occurance. Upon reaching 10000 the 0000
slot will be used again.

This change will firstly save the game, then copy the resulting slot to
a new autosave directory for the original slot.

Additionally, we will create new slot if playing on a loaded autosave to
avoid autosave-inception.

This should fix #1 since it's pretty much a complete rewrite of the
autosave feature.