refactor the build, make specific includes
Rather than using wildcards I'm going back to specific includes to avoid
situations where a new file was never added to source control for a
build. To this I'm also adding a clean step which removes the output
directroy in build/ - this to ensure a new build is "clean".

This commit will:
- remove duplicate QMods path
- make patch, .cs and language includes specific
- make clean remove the build output directory
- include settings.json on debug builds
- reset settings.json to sane values